requirements.txt
casadi==3.7.0
control==0.10.1
cvxpy==1.6.5
matlab==0.1
matplotlib==3.10.3
mediapy==1.2.4
mujoco==3.3.2
numpy==2.2.6
osR==0.0.1
pandas==2.3.0
plotly==6.1.2
PyYAML==6.0.2
scipy==1.15.3
setuptools==59.6.0
sympy==1.14.0
